kang_movie/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.4
django-filter==24.3
django-shortuuidfield==0.1.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.7
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
shortuuid==1.0.13
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
urllib3==2.2.2
